add docstrings and reorganize proj structure.
